Find top 10 clubs by average rating

+++ src/main/java/spark/mostpowerfulteam/Main.java
@@ -5,7 +5,7 @@ import org.apache.spark.api.java.JavaPairRDD;
 public class Main {
     public static void main(String[] args) {
         DataAnalyserService dataAnalyserService = new DataAnalyserService();
-        JavaPairRDD<String, Integer> teamToAggregatedOverallMapping = dataAnalyserService.findMostPowerfulTeam();
+        JavaPairRDD<String, Float> teamToAggregatedOverallMapping = dataAnalyserService.findMostPowerfulTeam();
 
         teamToAggregatedOverallMapping.take(10).forEach(System.out::println);
     }
